fix: Firefox CSP issue (#229)
* init

* add walletconnect script

* add vercel script

* add all possible headers

* remove few

* remove few more

* remove comments

* works on mozilla

* more updates

* move sha256 list to an array

* remove sha256

* tiny

* working with middleware

* middleware logic updated & split into different function

* remove nextSafe from config

* fix

* with headers

* update middelware

* returning the middleware response

* works

* add config matcher

* final changes
